ecommerce_backend system/requirements.txt
flask==3.0.0
mysql-connector-python==8.2.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
